+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 - SC-6511 - LDAP edit button missing.
 
+### Added
+
+- SC-6172: added hooks and checks to look for unique and not disposable emails in adminUsers service
+
 ## [24.4.0] - 2020-8-31
 
 ### Fixed
@@ -55,6 +59,10 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 - SC-5280: if disconnected prematurely, the LDAP service will not try to connect again just to unbind from the server
+
+### Added
+
+- SC-5808: added an isExternal check to the create method of AdminUsers service, only users from not external schools can create users
 
 ## [24.0.2] - 2020-08-05
 ### Fixed - 24.0.2